Safe.Partial

Description

ConstraintKind synonym for marking partial functions

Documentation

type Partial = HasCallStack

A constraint synonym which denotes that the function is partial, and will (on GHC 8.* and up) produce a stack trace on failure. You may mark your own non-total functions as Partial, if necessary, and this will ensure that they produce useful stack traces.